+++ _tools/.gitignore
@@ -1,0 +1,3 @@
+/dist/
+.env
+/node_modules/
