+++ _tools/README.md
@@ -1,0 +1,23 @@
+TOOLS  
+
+The main purpose for this tools is too scrap lot of game data from https://www.philibertnet.com
+
+1 . Scrapping data  
+
+        python3 scrapping.py
+
+This python script will output a games.json file containing the 245*36 html files.
+
+2. Fill database with the games.json  
+Once the games.json file is generated     
+launch npm scrpt
+    
+    
+    npm start
+
+        
+ You can choose what tables you want to fill by set the concerned CONSTANT to true on the fill_db.ts file.
+ 
+ For the database configuration, you should create a dot file .env to this folder.  
+ Check the README from app_server.
+
